requirements.txt
streamlit
flask
fastapi
uvicorn
langchain
langchain_community
langchain_core
psycopg2
openai-whisper
moviepy